backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis[asyncio]==5.0.1
httpx==0.25.2
pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.26.4
onnxruntime==1.19.0
